+++ .tape.js
@@ -1,0 +1,27 @@
+module.exports = {
+	'postcss-input-range': {
+		'basic': {
+			message: 'supports basic usage'
+		},
+		'basic:clone': {
+			message: 'supports "clone" method',
+			options: { method: 'clone' }
+		},
+		'basic:replace': {
+			message: 'supports "replace" method',
+			options: { method: 'replace' }
+		},
+		'basic:warn': {
+			message: 'supports "warn" method',
+			options: { method: 'warn' },
+			warning: 3
+		},
+		'vendor': {
+			message: 'ignores vendor prefixes if strict'
+		},
+		'vendor:loose': {
+			message: 'supports vendor prefixes if not strict',
+			options: { strict: false }
+		}
+	}
+};