+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-## 2.0.0 (2015-10-09)
+# Changes to Input Range
+
+### 3.0.0 (December 6, 2016)
+
+- Updated: boilerplate conventions (Node v6.9.1 LTS)
+
+### 2.0.0 (October 9, 2015)
 
 Added: `strict` option to process prefixed properties.
 Changed: Default `method` is `replace` as there is no spec.
 Updated: Documentation and tests
 
-## 1.0.0 (2015-10-09)
+### 1.0.0 (October 9, 2015)
 
 Added: Initial version